+++ src/HTTPCache.ts
@@ -300,7 +300,12 @@ function policyResponseFrom(response: FetcherResponse) {
   return {
     status: response.status,
     headers:
-      response.headers instanceof NodeFetchHeaders
+      response.headers instanceof NodeFetchHeaders &&
+      // https://github.com/apollo-server-integrations/apollo-server-integration-cloudflare-workers/issues/37
+      // For some reason, Cloudflare Workers' `response.headers` is passing
+      // the instanceof check here but doesn't have the `raw()` method that
+      // node-fetch's headers have.
+      'raw' in response.headers
         ? nodeFetchHeadersToCachePolicyHeaders(response.headers)
         : Object.fromEntries(response.headers),
   };
@@ -336,12 +341,12 @@ function cachePolicyHeadersToNodeFetchHeadersInit(
 ): NodeFetchHeadersInit {
   const headerList = [];
   for (const [name, value] of Object.entries(headers)) {
-    if (typeof value === 'string') {
-      headerList.push([name, value]);
-    } else if (value) {
+    if (Array.isArray(value)) {
       for (const subValue of value) {
         headerList.push([name, subValue]);
       }
+    } else if (value) {
+      headerList.push([name, value]);
     }
   }
   return headerList;
@@ -359,10 +364,10 @@ function cachePolicyHeadersToFetcherHeadersInit(
 ): Record<string, string> {
   const headerRecord = Object.create(null);
   for (const [name, value] of Object.entries(headers)) {
-    if (typeof value === 'string') {
-      headerRecord[name] = value;
-    } else if (value) {
+    if (Array.isArray(value)) {
       headerRecord[name] = value.join(', ');
+    } else if (value) {
+      headerRecord[name] = value;
     }
   }
   return headerRecord;